+++ Assets/Plugins/StreamChat/Core/DTO/Requests/UpdateChannelRequestDTO.cs
@@ -44,7 +44,7 @@ namespace StreamChat.Core.DTO.Requests
         /// Sets cool down period for the channel in seconds
         /// </summary>
         [Newtonsoft.Json.JsonProperty("cooldown", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
-        public double? Cooldown { get; set; }
+        public int? Cooldown { get; set; }
 
         /// <summary>
         /// Channel data to update
